(DRAFT) mupen64plus - rework logic including rpi aarch64 support and merging vero4k/mali  * temporarily disable rice and gles2n64 for aarch64 due to build errors (just HOST checking errors so may well be easily fixed)  * remove a lot of vero4k specific blocks in favour of mali flag  * enable mali for module (probably needs more work)
NOTE:
This is a combination of a WIP changeset I had, but adding aarch64 needed enough logic changes, it made sense to include it.
It may be best this is split out into its own PR, before aarch64 logic is added.

There has been a lot of contributions to this module, and unfortunately some of the logic has become rather obfuscated,
with a lot of duplication. I started work on trying to simplify it some time ago, including removal of the vero4k blocks,
but didn't finish so this should be considered incomplete as it is.